Make the signatures of methods consistent between headers and implementation files.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@29687 72102866-910b-0410-8b05-ffd578937521

+++ Headers/Additions/GNUstepGUI/GSLayoutManager.h
@@ -295,9 +295,9 @@ invalidation.)
 
 - (NSTextContainer *) textContainerForGlyphAtIndex: (NSUInteger)glyphIndex
 	effectiveRange: (NSRange *)effectiveRange;
-- (NSRect) lineFragmentRectForGlyphAtIndex: (unsigned int)glyphIndex
+- (NSRect) lineFragmentRectForGlyphAtIndex: (NSUInteger)glyphIndex
 	effectiveRange: (NSRange *)effectiveGlyphRange;
-- (NSRect) lineFragmentUsedRectForGlyphAtIndex: (unsigned int)glyphIndex
+- (NSRect) lineFragmentUsedRectForGlyphAtIndex: (NSUInteger)glyphIndex
 	effectiveRange: (NSRange *)effectiveGlyphRange;
 #if OS_API_VERSION(MAC_OS_X_VERSION_10_5, GS_API_LATEST)
 - (NSTextContainer *) textContainerForGlyphAtIndex: (NSUInteger)glyphIndex
